move endpoints to endpoints.py
