+++ components/matter/__init__.py
@@ -1,9 +1,7 @@
 from pathlib import Path
 
-from esphome import automation
 import esphome.codegen as cg
 import esphome.config_validation as cv
-from esphome.components import light, sensor
 from esphome.components.esp32 import (
     add_idf_component,
     add_idf_sdkconfig_option,
@@ -12,16 +10,17 @@ from esphome.components.esp32 import (
 from esphome.const import (
     CONF_ENABLE_IPV6,
     CONF_ID,
-    CONF_LIGHT_ID,
-    CONF_SENSOR_ID,
     Framework,
 )
 from esphome.core import CORE
 from esphome.coroutine import CoroPriority, coroutine_with_priority
 import esphome.final_validate as fv
 from esphome.helpers import write_file_if_changed
+from esphome.types import ConfigType
 
+from .endpoints import ENDPOINT_SCHEMA, configure_endpoints
 from .kconfig import disable_unused_clusters
+from .types import MatterComponent
 
 from .const import *
 
@@ -57,19 +56,6 @@ def _validate_passcode(value):
     return value
 
 
-matter_ns = cg.esphome_ns.namespace("matter")
-MatterComponent = matter_ns.class_("MatterComponent", cg.Component)
-MatterFactoryResetAction = matter_ns.class_(
-    "MatterFactoryResetAction", automation.Action
-)
-MatterEndpointRef = matter_ns.class_("MatterEndpointRef")
-MatterTurnOnAction = matter_ns.class_("MatterTurnOnAction", automation.Action)
-MatterTurnOffAction = matter_ns.class_("MatterTurnOffAction", automation.Action)
-MatterToggleAction = matter_ns.class_("MatterToggleAction", automation.Action)
-MatterDimAction = matter_ns.class_("MatterDimAction", automation.Action)
-MatterDimStopAction = matter_ns.class_("MatterDimStopAction", automation.Action)
-
-
 def _require_vfs_select(config):
     """Register VFS select requirement during config validation."""
     if CORE.is_esp32:
@@ -85,59 +71,6 @@ def _require_platformio_toolchain(config):
         )
     return config
 
-
-def _none_to_dict(value):
-    """Allow a bare `on_off_switch:` (no options)."""
-    return {} if value is None else value
-
-
-# Client switch endpoints take no options: they only define the Matter device
-# type (clusters + Binding). Behaviour is wired in YAML automations using the
-# matter.* actions, referencing the endpoint's id.
-ON_OFF_SWITCH_SCHEMA = cv.All(_none_to_dict, cv.Schema({}))
-
-DIMMER_SWITCH_SCHEMA = cv.All(_none_to_dict, cv.Schema({}))
-
-TEMPERATURE_SENSOR_SCHEMA = cv.Schema(
-    {
-        cv.Required(CONF_SENSOR_ID): cv.use_id(sensor.Sensor),
-    }
-)
-
-LIGHT_SCHEMA = cv.Schema(
-    {
-        cv.Required(CONF_LIGHT_ID): cv.use_id(light.LightState),
-    }
-)
-
-
-def _validate_endpoint(config):
-    device_types = [k for k in config if k != CONF_ID]
-    if len(device_types) != 1:
-        raise cv.Invalid(
-            "Each endpoint must have exactly one device type "
-            "(multiple device types per endpoint are not supported yet)"
-        )
-    return config
-
-
-# Each list entry is one Matter endpoint; the key selects the device type.
-# Endpoint ids are assigned in list order, so entries must never be removed
-# or reordered once the device is commissioned — append only.
-ENDPOINT_SCHEMA = cv.All(
-    cv.Schema(
-        {
-            # Referenceable from matter.* actions via endpoint_ref.
-            cv.GenerateID(): cv.declare_id(MatterEndpointRef),
-            cv.Optional(CONF_ON_OFF_SWITCH): ON_OFF_SWITCH_SCHEMA,
-            cv.Optional(CONF_DIMMER_SWITCH): DIMMER_SWITCH_SCHEMA,
-            cv.Optional(CONF_TEMPERATURE_SENSOR): TEMPERATURE_SENSOR_SCHEMA,
-            cv.Optional(CONF_ON_OFF_LIGHT): LIGHT_SCHEMA,
-            cv.Optional(CONF_DIMMABLE_LIGHT): LIGHT_SCHEMA,
-        }
-    ),
-    _validate_endpoint,
-)
 
 CONFIG_SCHEMA = cv.All(
     cv.Schema(
@@ -234,7 +167,7 @@ def _write_matter_cmake_hook() -> Path:
 
 
 @coroutine_with_priority(CoroPriority.NETWORK_SERVICES)
-async def to_code(config):
+async def to_code(config: ConfigType):
     var = cg.new_Pvariable(config[CONF_ID])
     await cg.register_component(var, config)
 
@@ -347,8 +280,6 @@ async def to_code(config):
         # TODO: set USE_BLE_ONLY_FOR_COMMISSIONING to false if other esphome components need it?
         # TODO: stop api from restarting device in commissioning mode?
 
-    disable_unused_clusters()
-
     # TODO: ENABLE_ESP32_FACTORY_DATA_PROVIDER?
 
     # CHIP's mbedTLS crypto backend calls mbedtls_hkdf() (CHIPCryptoPALmbedTLS.cpp).
@@ -367,93 +298,5 @@ async def to_code(config):
     # TODO: probably not needed?
     cg.add_build_flag("-DCHIP_CRYPTO_KEYSTORE_RAW=1")
 
-    for ep_conf in config[CONF_ENDPOINTS]:
-        ref = cg.new_Pvariable(ep_conf[CONF_ID])
-        if CONF_ON_OFF_SWITCH in ep_conf:
-            cg.add(var.add_on_off_switch(ref))
-        elif CONF_DIMMER_SWITCH in ep_conf:
-            cg.add(var.add_dimmer_switch(ref))
-        elif CONF_TEMPERATURE_SENSOR in ep_conf:
-            opts = ep_conf[CONF_TEMPERATURE_SENSOR]
-            sens = await cg.get_variable(opts[CONF_SENSOR_ID])
-            cg.add(var.add_temperature_sensor(sens, ref))
-        elif CONF_ON_OFF_LIGHT in ep_conf:
-            light_var = await cg.get_variable(ep_conf[CONF_ON_OFF_LIGHT][CONF_LIGHT_ID])
-            cg.add(var.add_on_off_light(light_var, ref))
-        elif CONF_DIMMABLE_LIGHT in ep_conf:
-            light_var = await cg.get_variable(
-                ep_conf[CONF_DIMMABLE_LIGHT][CONF_LIGHT_ID]
-            )
-            cg.add(var.add_dimmable_light(light_var, ref))
-
-
-@automation.register_action(
-    "matter.factory_reset",
-    MatterFactoryResetAction,
-    cv.Schema({cv.GenerateID(): cv.use_id(MatterComponent)}),
-)
-async def matter_factory_reset_to_code(config, action_id, template_arg, args):
-    var = cg.new_Pvariable(action_id, template_arg)
-    await cg.register_parented(var, config[CONF_ID])
-    return var
-
-
-# Accepts both `matter.turn_on: {id: my_endpoint}` and the short form
-# `matter.turn_on: my_endpoint`, like the light/switch component actions.
-MATTER_CLIENT_ACTION_SCHEMA = automation.maybe_simple_id(
-    {
-        cv.Required(CONF_ID): cv.use_id(MatterEndpointRef),
-    }
-)
-
-
-async def _matter_client_action_to_code(config, action_id, template_arg):
-    var = cg.new_Pvariable(action_id, template_arg)
-    await cg.register_parented(var, config[CONF_ID])
-    return var
-
-
-@automation.register_action(
-    "matter.turn_on", MatterTurnOnAction, MATTER_CLIENT_ACTION_SCHEMA
-)
-async def matter_turn_on_to_code(config, action_id, template_arg, args):
-    return await _matter_client_action_to_code(config, action_id, template_arg)
-
-
-@automation.register_action(
-    "matter.turn_off", MatterTurnOffAction, MATTER_CLIENT_ACTION_SCHEMA
-)
-async def matter_turn_off_to_code(config, action_id, template_arg, args):
-    return await _matter_client_action_to_code(config, action_id, template_arg)
-
-
-@automation.register_action(
-    "matter.toggle", MatterToggleAction, MATTER_CLIENT_ACTION_SCHEMA
-)
-async def matter_toggle_to_code(config, action_id, template_arg, args):
-    return await _matter_client_action_to_code(config, action_id, template_arg)
-
-
-@automation.register_action(
-    "matter.dim_up", MatterDimAction, MATTER_CLIENT_ACTION_SCHEMA
-)
-async def matter_dim_up_to_code(config, action_id, template_arg, args):
-    var = await _matter_client_action_to_code(config, action_id, template_arg)
-    cg.add(var.set_direction(0))
-    return var
-
-
-@automation.register_action(
-    "matter.dim_down", MatterDimAction, MATTER_CLIENT_ACTION_SCHEMA
-)
-async def matter_dim_down_to_code(config, action_id, template_arg, args):
-    var = await _matter_client_action_to_code(config, action_id, template_arg)
-    cg.add(var.set_direction(1))
-    return var
-
-
-@automation.register_action(
-    "matter.dim_stop", MatterDimStopAction, MATTER_CLIENT_ACTION_SCHEMA
-)
-async def matter_dim_stop_to_code(config, action_id, template_arg, args):
-    return await _matter_client_action_to_code(config, action_id, template_arg)
+    disable_unused_clusters()
+    await configure_endpoints(var, config)
